+++ test/Serilog.Sinks.Elasticsearch.Tests/Discrepancies/ElasticsearchDefaultSerializerTests.cs
@@ -1,5 +1,4 @@
-﻿using System.Linq;
-using Elasticsearch.Net.Serialization;
+﻿using Elasticsearch.Net;
 using NUnit.Framework;
 
 namespace Serilog.Sinks.Elasticsearch.Tests.Discrepancies