+++ test/Serilog.Sinks.Elasticsearch.Tests/Templating/SendsTemplateTests.cs
@@ -48,27 +48,14 @@ namespace Serilog.Sinks.Elasticsearch.Tests.Templating
 
         protected void JsonEquals(string json, MethodBase method, string fileName = null)
         {
-            var file = this.GetFileFromMethod(method, fileName);
-            var exists = File.Exists(file);
-            exists.Should().BeTrue(file + "does not exist");
-
-            var expected = File.ReadAllText(file);
+            var expected = TestDataHelper.ReadEmbeddedResource(Assembly.GetExecutingAssembly(), "template.json");
+            
             var nJson = JObject.Parse(json);
             var nOtherJson = JObject.Parse(expected);
             var equals = JToken.DeepEquals(nJson, nOtherJson);
             if (equals) return;
             expected.Should().BeEquivalentTo(json);
 
-        }
-        protected string GetFileFromMethod(MethodBase method, string fileName)
-        {
-            var type = method.DeclaringType;
-            var @namespace = method.DeclaringType.Namespace;
-            var folderSep = Path.DirectorySeparatorChar.ToString();
-            var folder = @namespace.Replace("Serilog.Sinks.Elasticsearch.Tests.", "").Replace(".", folderSep);
-            var file = Path.Combine(folder, (fileName ?? method.Name).Replace(@"\", folderSep) + ".json");
-            file = Path.Combine(Environment.CurrentDirectory.Replace("bin" + folderSep + "Debug", "").Replace("bin" + folderSep + "Release", ""), file);
-            return file;
         }
     }
 }